+++ app/src/main/java/com/example/myapplication/mainActivity/MainActivityModel.kt
@@ -5,11 +5,11 @@ import com.example.myapplication.database.weight.WeightsStorage
 class MainActivityModel(
     private val storage: WeightsStorage,
 ) {
-    fun getWeights(): List<Int> {
+    fun getWeights(): List<Float> {
         return storage.readWeights()
     }
 
-    fun addWeight(weight: Int): List<Int> {
+    fun addWeight(weight: Float): List<Float> {
         val updatedWeights = getWeights().toMutableList().apply {
             add(weight)
         }
@@ -17,7 +17,7 @@ class MainActivityModel(
         return updatedWeights
     }
 
-    fun removeWeight(index: Int): List<Int> {
+    fun removeWeight(index: Int): List<Float> {
         val updatedWeights = getWeights().toMutableList()
 
         if (index !in updatedWeights.indices) {
